Rename source files with duplicate names in different packages
Fixes #539
